refactor: 컨테이너 컴포넌트 활용
